Refactor solver modules into framework and greedy submodules
