+++ crates/dock-alloc-solver/src/framework/mod.rs
@@ -19,9 +19,5 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-pub mod berth;
-pub mod container;
-pub mod domain;
-pub mod framework;
-pub mod greedy;
-pub mod registry;
+pub mod planning;
+pub mod state;
