+++ crates/dock-alloc-solver/src/framework/planning.rs
